+++ disttask/framework/scheduler/interface_mock.go
@@ -99,7 +99,7 @@ func (m *MockPool) Run(f func()) error {
 }
 
 // RunWithConcurrency implements Pool.RunWithConcurrency.
-func (m *MockPool) RunWithConcurrency(funcs chan func(), _ uint64) error {
+func (m *MockPool) RunWithConcurrency(funcs chan func(), _ uint32) error {
 	args := m.Called()
 	if args.Error(0) == nil {
 		go func() {
